Refactor：will, retain
