+++ mqtt-meta/src/main/java/org/apache/rocketmq/mqtt/meta/raft/processor/WillMsgStateProcessor.java
@@ -17,22 +17,14 @@
 
 package org.apache.rocketmq.mqtt.meta.raft.processor;
 
-import com.alipay.sofa.jraft.util.BytesUtil;
+import org.apache.rocketmq.mqtt.common.meta.Constants;
 import org.apache.rocketmq.mqtt.common.model.consistency.ReadRequest;
 import org.apache.rocketmq.mqtt.common.model.consistency.Response;
 import org.apache.rocketmq.mqtt.common.model.consistency.WriteRequest;
 import org.apache.rocketmq.mqtt.meta.raft.MqttRaftServer;
 import org.apache.rocketmq.mqtt.meta.raft.MqttStateMachine;
-import org.apache.rocketmq.mqtt.common.meta.Constants;
-import org.apache.rocketmq.mqtt.meta.rocksdb.RocksDBEngine;
-import org.rocksdb.RocksIterator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.locks.Lock;
 
 import static org.apache.rocketmq.mqtt.common.meta.Constants.CATEGORY_WILL_MSG;
 
@@ -106,64 +98,6 @@ public class WillMsgStateProcessor extends StateProcessor {
     @Override
     public String groupCategory() {
         return CATEGORY_WILL_MSG;
-    }
-
-    public Response compareAndPut(RocksDBEngine rocksDBEngine, byte[] key, byte[] expectValue, byte[] updateValue) throws Exception {
-        final Lock writeLock = rocksDBEngine.getReadWriteLock().writeLock();
-        writeLock.lock();
-        try {
-            final byte[] actual = rocksDBEngine.getRdb().get(key);
-            if (Arrays.equals(expectValue, actual)) {
-                rocksDBEngine.getRdb().put(rocksDBEngine.getWriteOptions(), key, updateValue);
-                return Response.newBuilder()
-                        .setSuccess(true)
-                        .build();
-            } else {
-                return Response.newBuilder()
-                        .setSuccess(false)
-                        .build();
-            }
-        } catch (final Exception e) {
-            logger.error("Fail to delete, k {}", key, e);
-            throw e;
-        } finally {
-            writeLock.unlock();
-        }
-    }
-
-    public Response scan(RocksDBEngine rocksDBEngine, byte[] startKey, byte[] endKey) throws Exception {
-        Map<String, String> result = new HashMap<>();
-
-        final Lock readLock = rocksDBEngine.getReadWriteLock().readLock();
-        readLock.lock();
-        try {
-
-            final RocksIterator it = rocksDBEngine.getRdb().newIterator();
-            if (startKey == null) {
-                it.seekToFirst();
-            } else {
-                it.seek(startKey);
-            }
-
-            while (it.isValid()) {
-                final byte[] key = it.key();
-                if (endKey != null && BytesUtil.compare(key, endKey) >= 0) {
-                    break;
-                }
-                result.put(new String(key), new String(it.value()));
-                it.next();
-            }
-
-            return Response.newBuilder()
-                    .setSuccess(true)
-                    .putAllDataMap(result)
-                    .build();
-        } catch (final Exception e) {
-            logger.error("Fail to delete, startKey {}, endKey {}", startKey, endKey, e);
-            throw e;
-        } finally {
-            readLock.unlock();
-        }
     }
 
 }
